feat: 实现A-GPS功能与卫星历史记录
新增A-GPS数据管理功能，包括自动下载、手动注入和状态监控
添加卫星历史记录功能，支持保存和查看历史快照
重构GNSS数据模型，增加位置信息和卫星详细数据
扩展UI组件库，新增信号图表和历史卡片等组件
更新构建配置以支持数据序列化和WorkManager

+++ app/src/main/java/com/example/gpstest/domain/model/GnssSatellite.kt
@@ -10,8 +10,8 @@ data class GnssSatellite(
     val hasEphemeris: Boolean,
     val usedInFix: Boolean,
     val carrierFrequencyHz: Float?,
-    val carrierCycles: Float?,
-    val dopplerShiftHz: Float?,
+    val carrierCycles: Long?,
+    val dopplerShiftHz: Double?,
     val timeNanos: Long
 ) {
     val group: SatelliteGroup
